Enable test for macOS that seems to be working now
